Revert "Step22: ユーザの管理画面作成"
