+++ app/graphql/types/user_type.rb
@@ -4,10 +4,5 @@ module Types
     field :name, String, null: false
     field :created_at, GraphQL::Types::ISO8601DateTime, null: false
     field :updated_at, GraphQL::Types::ISO8601DateTime, null: false
-    field :tasks, [Types::TaskType], null: false
-    field :tasks_count, Int, null: false
-    def tasks_count
-      object.tasks.size
-    end
   end
 end
