+++ app/javascript/components/MyRouter.tsx
@@ -8,10 +8,6 @@ import TaskEdit from 'page/TaskEdit';
 import SignIn from 'page/SignIn';
 import SignUp from 'page/SignUp';
 import UnSignedTop from 'page/UnSignedTop';
-import AdminTop from 'page/AdminTop';
-import AdminUserCreate from 'page/AdminUserCreate';
-import AdminUserUpdate from 'page/AdminUserUpdate';
-import AdminUserTasks from 'page/AdminUserTasks';
 
 const MyRouter = () => {
   const { isLogin } = useContext(UserContext);
@@ -24,10 +20,6 @@ const MyRouter = () => {
         <Route path="/tasks/:id/edit" element={<TaskEdit />} />
         <Route path="/users/sign_in" element={<SignIn />} />
         <Route path="/users/sign_up" element={<SignUp />} />
-        <Route path="/admin" element={<AdminTop />} />
-        <Route path="/admin/users/new" element={<AdminUserCreate />} />
-        <Route path="/admin/users/:id/edit" element={<AdminUserUpdate />} />
-        <Route path="/admin/users/:id/tasks" element={<AdminUserTasks />} />
       </Routes>
     </BrowserRouter>
   );
