+++ app/javascript/components/utils/gql.tsx
@@ -18,7 +18,6 @@ const TASK_FRAGMENT = gql`
 export const GQL_TASKS = gql`
   ${TASK_FRAGMENT}
   query tasks(
-    $userId: ID
     $page: Int
     $word: String
     $doneIds: [ID!]
@@ -27,7 +26,6 @@ export const GQL_TASKS = gql`
     $target: String
   ) {
     tasks(
-      userId: $userId
       page: $page
       word: $word
       doneIds: $doneIds
@@ -116,47 +114,38 @@ export const GQL_DELETE_TASK = gql`
   }
 `;
 
-const USER_FRAGMENT = gql`
-  fragment UserFragment on User {
-    id
-    name
-    createdAt
-  }
-`;
-
 export const GQL_USER_SIGNED_IN = gql`
-  ${USER_FRAGMENT}
   {
     userSignedIn {
-      ...UserFragment
+      id
+      name
     }
   }
 `;
 
 export const GQL_SIGN_IN = gql`
-  ${USER_FRAGMENT}
   mutation signIn($name: String!, $password: String!) {
     signIn(input: { name: $name, password: $password }) {
       user {
-        ...UserFragment
+        id
+        name
       }
     }
   }
 `;
 
 export const GQL_SIGN_OUT = gql`
-  ${USER_FRAGMENT}
   mutation {
     signOut(input: {}) {
       user {
-        ...UserFragment
+        id
+        name
       }
     }
   }
 `;
 
 export const GQL_CREATE_USER = gql`
-  ${USER_FRAGMENT}
   mutation createUser(
     $name: String!
     $password: String!
@@ -170,94 +159,19 @@ export const GQL_CREATE_USER = gql`
       }
     ) {
       user {
-        ...UserFragment
+        id
+        name
       }
     }
   }
 `;
 
 export const GQL_DELETE_USER = gql`
-  ${USER_FRAGMENT}
   mutation {
     deleteUser(input: {}) {
       user {
-        ...UserFragment
-      }
-    }
-  }
-`;
-
-// for admin -------------------------------
-
-export const GQL_USER = gql`
-  ${USER_FRAGMENT}
-  query user($id: ID!) {
-    user(id: $id) {
-      ...UserFragment
-    }
-  }
-`;
-
-export const GQL_USERS = gql`
-  ${USER_FRAGMENT}
-  {
-    users {
-      ...UserFragment
-      tasksCount
-    }
-  }
-`;
-
-export const GQL_ADMIN_CREATE_USERS = gql`
-  ${USER_FRAGMENT}
-  mutation adminCreateUser(
-    $name: String!
-    $password: String!
-    $passwordConfirmation: String!
-  ) {
-    adminCreateUser(
-      input: {
-        name: $name
-        password: $password
-        passwordConfirmation: $passwordConfirmation
-      }
-    ) {
-      user {
-        ...UserFragment
-      }
-    }
-  }
-`;
-
-export const GQL_ADMIN_UPDATE_USER = gql`
-  ${USER_FRAGMENT}
-  mutation adminUpdateUser(
-    $id: ID!
-    $name: String
-    $password: String
-    $passwordConfirmation: String
-  ) {
-    adminUpdateUser(
-      input: {
-        id: $id
-        name: $name
-        password: $password
-        passwordConfirmation: $passwordConfirmation
-      }
-    ) {
-      user {
-        ...UserFragment
-      }
-    }
-  }
-`;
-
-export const GQL_ADMIN_DELETE_USER = gql`
-  ${USER_FRAGMENT}
-  mutation adminDeleteUser($id: ID!) {
-    adminDeleteUser(input: { id: $id }) {
-      user {
-        ...UserFragment
+        id
+        name
       }
     }
   }
